ai-learning-assistant/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
werkzeug==2.3.7
python-dotenv==1.0.0
google-generativeai==0.3.2
pymongo==4.6.1
mongoengine==0.27.0
flask-mail==0.9.1
